[Chore][Workloads] merge linear_attention family workloads
workloads/{deltanet,gla,gated_deltanet}.py -> workloads/linear_attention.py,
mirroring the manifest family. Bodies unchanged; 12 consumer imports
retargeted; collected node set byte-identical.
