+++ tests/ops/test_deltanet_fwd.py
@@ -4,7 +4,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import DeltaNetFwdOp
-from workloads.deltanet import DeltaNetFwdTest as _DeltaNetFwdTestWorkload
+from workloads.linear_attention import DeltaNetFwdTest as _DeltaNetFwdTestWorkload
 
 
 def compute_w_u_torch(Aw, Au, k, v, beta, chunk_size):
