+++ tests/ops/test_deltanet_recurrence.py
@@ -8,7 +8,7 @@ from tests.test_base import FixtureBase, TestBase
 from tileops.kernels.deltanet_recurrence import DeltaNetDecodeRawCudaFlaStyleKernel
 from tileops.kernels.kernel_base import Kernel
 from tileops.ops import DeltaNetDecodeOp
-from workloads.deltanet import DeltaNetDecodeTest as _DeltaNetDecodeTestWorkload
+from workloads.linear_attention import DeltaNetDecodeTest as _DeltaNetDecodeTestWorkload
 
 
 def deltanet_decode_torch(
