+++ tests/ops/test_gla_recurrence.py
@@ -5,7 +5,7 @@ import torch
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.ops import GLADecodeOp
-from workloads.gla import GLADecodeTest as _GLADecodeTestWorkload
+from workloads.linear_attention import GLADecodeTest as _GLADecodeTestWorkload
 
 
 def gla_decode_torch(
